Fixed memory issue.
FINALLY! The memory issue is actually a combination of couple of problems. First, WinRT itself does not like recursive data templates. Second, I never actually unsubscribed any of the custom event handlers I put the ObservableCollections. The event handlers causes circular reference between xaml and the viewmodel which GC is not very good at cleaning up. That ramps up the memory usage every time we reassign the list. Third, we are not really supposed to be reassigning lists. When xaml binds to an ObservableCollection, we should try as much as we can to keep that instance and not abuse property changed event. The infinite memory ramp is probably caused by a combination of circular event reference and 'dead' ObservableCollections.

To solve the above issues I decided to rework how comments are displayed. The comments still uses a tree as its core data structure but now it is falttened into a list before feeding it to the UI. This avoid recursive datatemplates and reducing the number of ObservableCollection to just one for every SubmissionCommentsVM. Even though the comments are fed to the UI as a list I still managed to give it a tree like handling and look. The visibility toggling of the comments in the list is achieved through two visibility property for each comment instead of one. A backstack of the visibility is kept to save state information between toggling of top level comments. Pretty clever if I do say so myself.